Worked on integrating invetory system into player attack system
