+++ items/__init__.py
@@ -1,2 +1,3 @@
 from .item import Item
+from .dagger import Dagger
 from .sword import Sword
