+++ items/dagger.py
@@ -2,20 +2,20 @@ from .item import Item
 from inventory import InventoryItem
 from time import time
 
-class Sword(Item):
-    """Represents the base sword"""
+class Dagger(Item):
+    """Represents the base dagger"""
     def __init__(self, owner):
         super().__init__(owner, surpressIICreation = True)
-        self.damage = 1
-        self.delay = 0.5
+        self.damage = 0.2
+        self.delay = 0.25
         self.lastUse = -1
         hasImage = True
         self.inventoryItem = InventoryItem(
             self.owner.inventory,
-            "Sword",
-            groups = ["Weapon", "Sword"],
+            "Dagger",
+            groups = ["Weapon", "Dagger"],
             category = "Weapon",
-            description = "Base Sword",
+            description = "Base Dagger",
             owners = [self],
             stats = {
                 "attackDamage": 5
